+++ CHANGELOG.md
@@ -3,6 +3,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Forge Recommended Versioning](https://mcforge.readthedocs.io/en/latest/conventions/versioning/).
 
+## [1.19.2-3.0.2.0] - 2022-10-05
+### Added
+- Lucent mod support for Mining Hat (if Lucent is installed, the night vision effect is replaced by the Lucent light effect)
+
+### Changed
+- optimize Mining Hat damaging to be more precisely
+
 ## [1.19.2-3.0.1.0] - 2022-09-13
 ### Changed
 - Update mod to Forge 1.19.2-43.1.23